+++ src/ThumbCreator.php
@@ -20,7 +20,6 @@ use Intervention\Image\Image;
 use Intervention\Image\ImageManager;
 use LogicException;
 use Symfony\Component\Filesystem\Exception\IOException;
-use Thumber\Exception\NotReadableImageException;
 use Thumber\Exception\UnsupportedImageTypeException;
 use Tools\Filesystem;
 
@@ -111,8 +110,6 @@ class ThumbCreator
     /**
      * Gets an `Image` instance
      * @return \Intervention\Image\Image
-     * @throws \Thumber\Exception\NotReadableImageException
-     * @throws \Thumber\Exception\UnsupportedImageTypeException
      * @throws \ErrorException
      */
     protected function getImageInstance(): Image
@@ -125,7 +122,7 @@ class ThumbCreator
                 throw new UnsupportedImageTypeException($type ? 'Image type `' . $type . '` is not supported by this driver' : 'Image type not supported by this driver');
             }
             $path = $this->Filesystem->rtr($this->path) ?: null;
-            throw new NotReadableImageException($path ? 'Unable to read image from `' . $path . '`' : 'Unable to read image from file');
+            throw new LogicException($path ? 'Unable to read image from `' . $path . '`' : 'Unable to read image from file');
         }
 
         return $ImageInstance;