+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.5 branch
 ### 1.5.4
 * updated for `php-tools` 1.8;
+* `NotReadableImageException` no longer exists, replaced with a generic `LogicException`;
 * no longer uses the `Exceptionist` class.
 
 ### 1.5.3